fix: symbol check in older browsers
